feat: add flag to hide menus unneeded for pre-event
